+++ src/components/email-confirm-modal/emailConfirmationModal.style.js
